requirements.txt
accelerate==0.24.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.11
boto3==1.28.73
botocore==1.31.73
bpemb==0.3.4
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
conllu==4.5.3
contourpy==1.1.1
coreschema==0.0.4
cycler==0.12.1
cymem==2.0.8
Deprecated==1.2.14
Django==4.2.6
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
emoji==2.8.0
filelock==3.13.0
flair==0.13.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.1
gdown==4.7.1
gensim==4.3.2
huggingface-hub==0.17.3
idna==3.4
inflection==0.5.1
Janome==0.5.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
lemminflect==0.2.3
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==10.1.0
mpld3==0.5.9
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
openai==0.28.1
packaging==23.2
Pillow==10.1.0
pptree==3.1
preshed==3.0.9
protobuf==4.24.4
psutil==5.9.6
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch_revgrad==0.2.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
s3transfer==0.7.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
segtok==1.5.11
semver==3.0.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-stanza==1.0.4
sqlitedict==2.1.0
sqlparse==0.4.4
srsly==2.4.8
stanza==1.6.1
sympy==1.12
tabulate==0.9.0
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformer-smaller-training-vocab==0.3.2
transformers==4.34.1
triton==2.1.0
typer==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.18
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.3
Wikipedia-API==0.6.0
wrapt==1.15.0
yarl==1.9.2
